REF: move a cumulative method that slipped through the cracks
