+++ pandas/tests/series/test_cumulative.py
@@ -5,6 +5,8 @@ See also
 --------
 tests.frame.test_cumulative
 """
+from itertools import product
+
 import numpy as np
 import pytest
 
@@ -140,3 +142,33 @@ class TestSeriesCumulativeOps:
         )
         result = s.cummax(skipna=False)
         tm.assert_series_equal(expected, result)
+
+    def test_cummethods_bool(self):
+        # GH#6270
+
+        a = pd.Series([False, False, False, True, True, False, False])
+        b = ~a
+        c = pd.Series([False] * len(b))
+        d = ~c
+        methods = {
+            "cumsum": np.cumsum,
+            "cumprod": np.cumprod,
+            "cummin": np.minimum.accumulate,
+            "cummax": np.maximum.accumulate,
+        }
+        args = product((a, b, c, d), methods)
+        for s, method in args:
+            expected = pd.Series(methods[method](s.values))
+            result = getattr(s, method)()
+            tm.assert_series_equal(result, expected)
+
+        e = pd.Series([False, True, np.nan, False])
+        cse = pd.Series([0, 1, np.nan, 1], dtype=object)
+        cpe = pd.Series([False, 0, np.nan, 0])
+        cmin = pd.Series([False, False, np.nan, False])
+        cmax = pd.Series([False, True, np.nan, True])
+        expecteds = {"cumsum": cse, "cumprod": cpe, "cummin": cmin, "cummax": cmax}
+
+        for method in methods:
+            res = getattr(e, method)()
+            tm.assert_series_equal(res, expecteds[method])
